Move test result files to a new place.
